reorganize archive api
