+++ include/Pothos/Archive.hpp
@@ -1,0 +1,14 @@
+///
+/// \file Pothos/Archive.hpp
+///
+/// Top level include wrapper for Archive classes.
+///
+/// \copyright
+/// Copyright (c) 2016 Josh Blum
+/// SPDX-License-Identifier: BSL-1.0
+///
+
+#pragma once
+#include <Pothos/Config.hpp>
+#include <Pothos/Archive/Archive.hpp>
+#include <Pothos/Archive/Macros.hpp>
