+++ lib/Archive/Archive.cpp
@@ -1,10 +1,12 @@
 // Copyright (c) 2016 Josh Blum
 // SPDX-License-Identifier: BSL-1.0
 
-#include <Pothos/Util/Archive.hpp>
+#include <Pothos/Archive.hpp>
+#include <Pothos/Archive/String.hpp>
+#include <Pothos/Archive/Numbers.hpp>
 #include <iostream>
 
-Pothos::Util::ArchiveEntry::ArchiveEntry(const std::type_info &type, const char *id)
+Pothos::Archive::ArchiveEntry::ArchiveEntry(const std::type_info &type, const char *id)
 {
     std::cout << "REGISTER " << type.name() << ", id = " << id << std::endl;
 }
@@ -41,5 +43,5 @@ void test(void)
 {
     BooHoo bh;
     std::stringstream ss;
-    Pothos::Util::SerializeArchive(ss, bh);
+    Pothos::Archive::serializeArchive(ss, bh);
 }
